requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
redis>=4.0,<5.0
djangorestframework>=3.15.2
celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.7.0
openai==1.58.1
python-dotenv==1.0.1
django-cors-headers==4.6.0
